apps/backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.30.6
psycopg[binary,pool]==3.2.3
pgvector==0.2.5
sentence-transformers==3.0.1
python-multipart==0.0.9
pdfplumber==0.11.4
python-docx==1.1.2
openai==1.54.4
